refactor(billing): reduce the Stripe webhook handler to a thin dispatcher
Move the settlement, refund and payment-method-sync orchestration out of the webhook and
into the owner services. StripeTransactionService gains event-facing settlePaymentIntent,
settleInvoice, refundCharge, failPaymentIntent and cancelPaymentIntent, resolving the user
and enriching charge details itself. PaymentMethodService gains syncAttachedFromEvent and
removeDetachedFromEvent.

StripeWebhookService now only verifies the signature, dispatches to those owners, and
publishes the first-purchase-bonus domain event. It no longer depends on UserRepository or
StripeTransactionRepository and touches no records directly.

The settlement and refund tests move to the owner services as integration specs (they cross
@WithTransaction writers); the webhook keeps a unit dispatch spec. No behavior change.
